Draw black or white pawns or rooks

+++ Chess/PieceSetter.cpp
@@ -4,7 +4,7 @@ void PieceSetter::sendPiecesToBoard(Positions p, Board &b){
 	/*
 		Update squares of board to have the pieces described in positions
 	*/
-	for(int i = 0; i < p.getPieces().size(); i++){
+	for(size_t i = 0; i < p.getPieces().size(); i++){
 		b.update(p.getPieces()[i]);	
 	}	
 }
